Merging develop to master

+++ server/main.go
@@ -4,28 +4,29 @@ import (
 	"context"
 	"fmt"
 	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
+	"github.com/spf13/viper"
 	"google.golang.org/grpc"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/types/known/timestamppb"
 	"log"
 	"net"
 	"net/http"
 	"os"
+	"regexp"
 	db "server/main.go/database"
 	models "server/main.go/models"
 	userpb "server/main.go/proto"
 )
 
 const (
-	address     = "localhost:8080"
-	colorRed    = "\033[31m"
 	colorGreen  = "\033[32m"
-	colorWhite  = "\033[37m"
 	colorPurple = "\033[35m"
 	colorBlue   = "\033[34m"
 	colorYellow = "\033[33m"
 )
 
-var database = db.Connect()
+var database = db.Connect().Debug()
 
 type userServer struct {
 	userpb.UnimplementedUserServiceServer
@@ -33,6 +34,30 @@ type userServer struct {
 
 type creditCardServer struct {
 	userpb.UnimplementedCreditCardServiceServer
+}
+
+func allowedOrigin(origin string) bool {
+	if viper.GetString("cors") == "*" {
+		return true
+	}
+	if matched, _ := regexp.MatchString(viper.GetString("cors"), origin); matched {
+		return true
+	}
+	return false
+}
+
+func cors(h http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if allowedOrigin(r.Header.Get("Origin")) {
+			w.Header().Set("Access-Control-Allow-Origin", r.Header.Get("Origin"))
+			w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PATCH, DELETE")
+			w.Header().Set("Access-Control-Allow-Headers", "Accept, Content-Type, Content-Length, Accept-Encoding, Authorization, ResponseType")
+		}
+		if r.Method == "OPTIONS" {
+			return
+		}
+		h.ServeHTTP(w, r)
+	})
 }
 
 func (as *userServer) SayHello(ctx context.Context, in *userpb.User) (*userpb.Message, error) {
@@ -51,6 +76,9 @@ func (as *userServer) CreateUser(ctx context.Context, in *userpb.User) (*userpb.
 func (as *userServer) UpdateUserByName(ctx context.Context, in *userpb.User) (*userpb.User, error) {
 
 	name := in.GetName()
+	if name == "" {
+		return &userpb.User{}, status.Error(codes.InvalidArgument, "User's name not specified")
+	}
 	email := in.GetEmail()
 	phoneNumber := in.GetPhoneNumber()
 
@@ -86,8 +114,15 @@ func (as *userServer) UpdateUserById(ctx context.Context, in *userpb.User) (*use
 
 func (as *userServer) DeleteUser(ctx context.Context, in *userpb.User) (*userpb.Message, error) {
 	name := in.GetName()
+	if name == "" {
+		return &userpb.Message{}, status.Error(codes.InvalidArgument, "User's name not specified")
+	}
 	var user models.User
-	database.Where("name =?", name).Find(&user)
+	rowsAffected := database.Where("name =?", name).Find(&user).RowsAffected
+
+	if rowsAffected == 0 {
+		return &userpb.Message{}, status.Error(codes.NotFound, "Cannot find a User with this name!")
+	}
 	database.Delete(&user)
 
 	return &userpb.Message{Message: user.Name + " Deleted successfully!"}, nil
@@ -96,7 +131,7 @@ func (as *userServer) DeleteUser(ctx context.Context, in *userpb.User) (*userpb.
 func (as *userServer) ListUsers(ctx context.Context, in *userpb.User) (*userpb.ListUser, error) {
 
 	list := make([]*userpb.User, 0)
-	database.Where("deleted_at is null").Order("created_at desc").Find(&list)
+	database.Where("deleted_at is null").Order("created_at desc").Limit(100).Find(&list)
 	return &userpb.ListUser{
 		Users: list,
 	}, nil
@@ -111,6 +146,9 @@ func (as *userServer) GetUserByName(ctx context.Context, in *userpb.User) (*user
 	fmt.Println(colorYellow, "__________________________________________________________________________________"+
 		"_______________________________________________________________________________________________________")
 	name := in.GetName()
+	if name == "" {
+		return &userpb.User{}, status.Error(codes.InvalidArgument, "User's name not specified")
+	}
 	var user models.User
 	database.Where(&models.User{Name: name}).Find(&user)
 
@@ -120,7 +158,11 @@ func (as *userServer) GetUserByName(ctx context.Context, in *userpb.User) (*user
 func (as *userServer) GetUserById(ctx context.Context, in *userpb.User) (*userpb.User, error) {
 	id := in.GetId()
 	var user models.User
-	database.Where("id = ?", id).Find(&user)
+	rowsAffected := database.Where("id = ?", id).Find(&user).RowsAffected
+
+	if rowsAffected == 0 {
+		return &userpb.User{}, status.Error(codes.NotFound, "Cannot find a User with this id!")
+	}
 
 	return &userpb.User{Id: uint32(user.ID), Name: user.Name, Email: user.Email, PhoneNumber: user.PhoneNumber}, nil
 }
@@ -171,14 +213,114 @@ func (fu *creditCardServer) GetCreditCardByUserName(ctx context.Context, in *use
 	log.SetFlags(log.LstdFlags | log.Lmicroseconds)
 
 	name := in.GetName()
+	if name == "" {
+		return &userpb.CreditCard{}, status.Error(codes.InvalidArgument, "User's name cannot be empty")
+	}
 	var creditCard models.CreditCards
-	database.Where(&models.CreditCards{Name: name}).Find(&creditCard)
+	rowsAffected := database.Where(&models.CreditCards{Name: name}).Find(&creditCard).RowsAffected
+
+	if rowsAffected == 0 {
+		return &userpb.CreditCard{}, status.Error(codes.NotFound, "Cannot find a credit card with this user name!")
+	}
 
 	log.Println(colorPurple, "[creditCardService] - [rpc GetCreditCardByUserName] -> ", colorGreen, "Now sending the response (credit card of selected user) to client side...")
 
 	return &userpb.CreditCard{Id: uint32(creditCard.ID), Name: creditCard.Name, Email: creditCard.Email, PhoneNumber: creditCard.PhoneNumber,
 		Address: creditCard.Address, Country: creditCard.Country, City: creditCard.City, Zip: creditCard.Zip, Cvv: creditCard.CVV,
 		CreatedAt: timestamppb.New(creditCard.CreatedAt)}, nil
+}
+
+func (fu *creditCardServer) CreateCreditCardApplication(ctx context.Context, in *userpb.CreditCardApplication) (*userpb.CreditCardApplication, error) {
+
+	creditCardApplication := models.CreditCardApplication{
+		FirstName:            in.GetFirstName(),
+		LastName:             in.GetLastName(),
+		DateOfBirth:          in.GetDateOfBirth().AsTime(),
+		PhoneNumber:          in.GetPhoneNumber(),
+		SocialSecurityNumber: in.GetSocialSecurityNumber(),
+		EmploymentType:       in.GetEmploymentType(),
+		Occupation:           in.GetOccupation(),
+		MonthlyIncome:        float64(in.GetMonthlyIncome()),
+		YearsEmployed:        int(in.GetYearsEmployed()),
+		StreetAddress:        in.GetStreetAddress(),
+		YearsAtAddress:       int(in.GetYearsAtAddress()),
+		City:                 in.GetCity(),
+		State:                in.GetState(),
+		Zip:                  in.GetZip(),
+		Country:              in.GetCountry(),
+		Ownership:            in.GetOwnership(),
+		MonthlyPayment:       float64(in.GetMonthlyPayment()),
+		CardName:             in.GetCardName(),
+		CardType:             in.GetCardType(),
+		Branch:               in.GetBranch(),
+		CardBranding:         in.GetCardBranding(),
+	}
+
+	database.NewRecord(creditCardApplication)
+	database.Create(&creditCardApplication)
+
+	return &userpb.CreditCardApplication{
+		Id:                   uint32(creditCardApplication.ID),
+		FirstName:            creditCardApplication.FirstName,
+		LastName:             creditCardApplication.LastName,
+		DateOfBirth:          timestamppb.New(creditCardApplication.DateOfBirth),
+		PhoneNumber:          creditCardApplication.PhoneNumber,
+		SocialSecurityNumber: creditCardApplication.SocialSecurityNumber,
+		EmploymentType:       creditCardApplication.EmploymentType,
+		Occupation:           creditCardApplication.Occupation,
+		MonthlyIncome:        float32(creditCardApplication.MonthlyIncome),
+		YearsEmployed:        int32(creditCardApplication.YearsEmployed),
+		StreetAddress:        creditCardApplication.StreetAddress,
+		YearsAtAddress:       int32(creditCardApplication.YearsAtAddress),
+		City:                 creditCardApplication.City,
+		State:                creditCardApplication.State,
+		Zip:                  creditCardApplication.Zip,
+		Country:              creditCardApplication.Country,
+		Ownership:            creditCardApplication.Ownership,
+		MonthlyPayment:       float32(creditCardApplication.MonthlyPayment),
+		CardName:             creditCardApplication.CardName,
+		CardType:             creditCardApplication.CardType,
+		Branch:               creditCardApplication.Branch,
+		CardBranding:         creditCardApplication.CardBranding,
+		CreatedAt:            timestamppb.New(creditCardApplication.CreatedAt),
+		UpdatedAt:            timestamppb.New(creditCardApplication.UpdatedAt),
+		DeletedAt:            timestamppb.New(creditCardApplication.DeletedAt),
+	}, nil
+}
+
+func (fu *creditCardServer) GetCreditCardApplicationByName(ctx context.Context, in *userpb.CreditCardApplication) (*userpb.CreditCardApplication, error) {
+
+	firstName := in.GetFirstName()
+	var creditCardApplication models.CreditCardApplication
+	database.Unscoped().Where(&models.CreditCardApplication{FirstName: firstName}).Order("created_at desc").First(&creditCardApplication)
+
+	return &userpb.CreditCardApplication{
+		Id:                   uint32(creditCardApplication.ID),
+		FirstName:            creditCardApplication.FirstName,
+		LastName:             creditCardApplication.LastName,
+		DateOfBirth:          timestamppb.New(creditCardApplication.DateOfBirth),
+		PhoneNumber:          creditCardApplication.PhoneNumber,
+		SocialSecurityNumber: creditCardApplication.SocialSecurityNumber,
+		EmploymentType:       creditCardApplication.EmploymentType,
+		Occupation:           creditCardApplication.Occupation,
+		MonthlyIncome:        float32(creditCardApplication.MonthlyIncome),
+		YearsEmployed:        int32(creditCardApplication.YearsEmployed),
+		StreetAddress:        creditCardApplication.StreetAddress,
+		YearsAtAddress:       int32(creditCardApplication.YearsAtAddress),
+		City:                 creditCardApplication.City,
+		State:                creditCardApplication.State,
+		Zip:                  creditCardApplication.Zip,
+		Country:              creditCardApplication.Country,
+		Ownership:            creditCardApplication.Ownership,
+		MonthlyPayment:       float32(creditCardApplication.MonthlyPayment),
+		CardName:             creditCardApplication.CardName,
+		CardType:             creditCardApplication.CardType,
+		Branch:               creditCardApplication.Branch,
+		CardBranding:         creditCardApplication.CardBranding,
+		CreatedAt:            timestamppb.New(creditCardApplication.CreatedAt),
+		UpdatedAt:            timestamppb.New(creditCardApplication.UpdatedAt),
+		DeletedAt:            timestamppb.New(creditCardApplication.DeletedAt),
+	}, nil
 }
 
 func main() {
@@ -227,7 +369,7 @@ func main() {
 
 	gwServer := &http.Server{
 		Addr:    fmt.Sprintf(":%s", os.Getenv("server_port")),
-		Handler: gwmux,
+		Handler: cors(gwmux),
 	}
 
 	log.Println(fmt.Sprintf("Serving gRPC-Gateway on %s:%s", os.Getenv("server_host"), os.Getenv("server_port")))
